SMP-12: Added in Google authentication and ability to sign in and out

+++ app/src/main/java/com/zonkey/simplemealplanner/activity/RecipeDetailActivityPresenter.kt
@@ -11,17 +11,21 @@ class RecipeDetailActivityPresenter(
     private val firebaseRepo: FirebaseRecipeRepository
 ) {
 
-  fun onFavoriteButtonClicked(savedRecipe: Boolean, recipe: Recipe) {
-    if (savedRecipe) {
-      firebaseRepo.removeRecipeAsFavorite(recipe)
-      view.isSavedRecipe = false
-      setSavedRecipeIcon(false)
-      view.showRecipeDetailSnackBar(R.string.snackbar_recipe_deleted)
+  fun onFavoriteButtonClicked(isSignedIn: Boolean, savedRecipe: Boolean, recipe: Recipe) {
+    if (!isSignedIn) {
+      view.launchUIAuthActivity()
     } else {
-      firebaseRepo.saveRecipeAsFavorite(recipe)
-      view.isSavedRecipe = true
-      setSavedRecipeIcon(true)
-      view.showRecipeDetailSnackBar(R.string.snackbar_recipe_saved)
+      if (savedRecipe) {
+        firebaseRepo.removeRecipeAsFavorite(recipe)
+        view.isSavedRecipe = false
+        setSavedRecipeIcon(false)
+        view.showRecipeDetailSnackBar(R.string.snackbar_recipe_deleted)
+      } else {
+        firebaseRepo.saveRecipeAsFavorite(recipe)
+        view.isSavedRecipe = true
+        setSavedRecipeIcon(true)
+        view.showRecipeDetailSnackBar(R.string.snackbar_recipe_saved)
+      }
     }
   }
 
@@ -33,29 +37,41 @@ class RecipeDetailActivityPresenter(
     }
   }
 
-  fun onMealPlanDialogPositiveButtonClicked(recipe: Recipe, addedToMealPlan: Boolean,
-      selectedDay: String, isSavedRecipe: Boolean) {
-    when {
-      addedToMealPlan -> firebaseRepo.updateMealPlanRecipeDay(recipe,
-          DayOfWeek.valueOf(selectedDay))
-      else -> {
-        firebaseRepo.saveRecipeToMealPlan(recipe, DayOfWeek.valueOf(selectedDay), isSavedRecipe)
-        view.addedToMealPlan = true
+  fun onMealPlanDialogPositiveButtonClicked(
+      isSignedIn: Boolean,
+      recipe: Recipe,
+      addedToMealPlan: Boolean,
+      selectedDay: String,
+      isSavedRecipe: Boolean) {
+    if (!isSignedIn) {
+      view.launchUIAuthActivity()
+    } else {
+      when {
+        addedToMealPlan -> firebaseRepo.updateMealPlanRecipeDay(recipe,
+            DayOfWeek.valueOf(selectedDay))
+        else -> {
+          firebaseRepo.saveRecipeToMealPlan(recipe, DayOfWeek.valueOf(selectedDay), isSavedRecipe)
+          view.addedToMealPlan = true
+        }
       }
-    }
-    when (DayOfWeek.valueOf(selectedDay)) {
-      REMOVE -> {
-        view.setMealPlanButtonText(mealPlanButtonStringRes = R.string.detail_meal_plan_button_text)
-        firebaseRepo.removeRecipeFromMealPlan(recipe)
+      when (DayOfWeek.valueOf(selectedDay)) {
+        REMOVE -> {
+          view.setMealPlanButtonText(
+              mealPlanButtonStringRes = R.string.detail_meal_plan_button_text)
+          firebaseRepo.removeRecipeFromMealPlan(recipe)
+        }
+        else -> view.setMealPlanButtonText(selectedDayString = selectedDay)
       }
-      else -> view.setMealPlanButtonText(selectedDayString = selectedDay)
+      showRecipeDetailSnackBar(selectedDay)
     }
   }
 
   fun showRecipeDetailSnackBar(selectedDay: String) {
     when (selectedDay) {
-      REMOVE.name -> view.showRecipeDetailSnackBar(snackBarStringRes = R.string.detail_snackbar_meal_plan_removed)
-      else -> view.showRecipeDetailSnackBar(snackBarStringRes = R.string.detail_meal_plan_snackbar_text, dayOfWeek = selectedDay)
+      REMOVE.name -> view.showRecipeDetailSnackBar(
+          snackBarStringRes = R.string.detail_snackbar_meal_plan_removed)
+      else -> view.showRecipeDetailSnackBar(
+          snackBarStringRes = R.string.detail_meal_plan_snackbar_text, dayOfWeek = selectedDay)
     }
   }
 }
